More consistently apply specific attribute resets (unformat)
c.unformat has logic for resetting specific attributes for the
instance of Color 'c'. This is useful when applying several Colors with
different attributes to a string or io stream, as Colors will aim to
only reset escape codes *they* applied, leaving others' intact.

c.format is invoked in 3 places, but c.unformat is invoked only in one
spot, highlighting an asymmetry. If we look at these missing spots, they
are:

 1. c.Set vs. c.unset
 2. c.SetWriter vs. c.UnsetWriter

It makes sense to apply the 'c.unformat' logic to these two cases. We
don't touch the global 'Unset' function as it does not have the context
of a particular Color, and so continues to apply a global reset code.

While here, we also make the private 'c.unset' into an exported
'c.Unset' so that 'c.Set' has a counterpart.

The added tests all fail prior to this commit, but pass after.