Add Hyperlink support
[1] lays out a standard for terminals to support hyperlinked text.
Support is quite widespread but not universal. Examples of support
include iTerm2, Ghostty, Alacritty, Hyper, Kitty, and more [2].

This OSC 8 standard is achieved via escape codes, very similar to what
this project does with colors and other attributes - it's a natural fit.

Because it is a little unique in how it functions, this 'hyperlink'
feature doesn't follow the existing 'Attribute' pattern. Instead, it's a
standalone field which gets special handling. That said, updating just
c.format and c.unformat is enough, as with the last commit, these are
now used in all relevant spots that affect formatting.

[1] https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
[2] https://github.com/Alhadis/OSC8-Adoption/?tab=readme-ov-file